Update to work with Composer at 0627d84
A change was made to composer that removes plugin loading on exception.
Since exploiting the "command not found" exception is no longer possible,
the completion command has to piggy-back on another command that does load
plugins and has a compatible signature.

Annoyingly, this means the completion command is no longer functional in
the same way as if it was included in Composer properly, but this is an
intentional hack, so you get what you pay for I guess.

Resolves #8

Commit that necessitated this update:
https://github.com/composer/composer/commit/0627d846a5a4b4378ad74851d877c3f5f34a6aaa