Update symfony-completion-command to 0.7.0
This updates the completion library to a version that can be installed
alongside Symfony Console 3.x. None of the other improvements in the library
benefit this plugin, but the hook scripts have had to be updated as part of
the upgrade (which is why I hadn't updated it for ages).

Fixes #9