Query Composer to determine global config dir
Composer has a decent chunk of logic for deciding which directory to use
as the global config dir. Rather than mirror that logic in shell scripts,
I've updated the hooks to use `composer config --global home` to
determine the config dir.

Running `composer config --global home` on my machine takes around 100ms
on average, so I've also updated the install instructions to generate a
fixed path for `source`'ing (sluggish shell profiles are a pain). The
config dir is also queried on the first completion attempt. I've done
this instead of adding an environment variable to the profile, as this
implementation is backwards compatible with existing installs.

Fixes #10